backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]
python-multipart==0.0.6
bcrypt==4.0.1
pyjwt==2.8.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.1.3